refactor: fix naming for view actions
